#49 公開ソースとの突き合わせ: 0x8002 がページ送りになる 7 ファイル / 表の位置は u32
- turi_info / phot_info / okan_info / item_info / insect_menu / fishing / fish_info の
  .msg では 0x8002 に引数が無い (MSG.py の ALT_NEWLINE_FILES)。Python / ブラウザとも
  ファイル名で見分けて <BREAK> として読み、次の値を文字として数える
- 表の一覧の項目の位置を u32 で読む (MSG_notes.txt)。64 KiB を超える位置の表のテスト
- 練習データに system/item_info.msg (<BREAK> 入り) を追加し、答え合わせに含める
- docs/09: 判明済みに追記、#49 を記録

+++ tests/test_fuzz.mjs
@@ -51,7 +51,7 @@ function buildTables(tables) {
   const dv = new DataView(buf.buffer);
   dv.setUint32(0, tables.length, true);
   let p = head;
-  bodies.forEach((body, i) => { dv.setUint32(4 + i * 12, 0xDEAD, true); dv.setUint16(8 + i * 12, body.length, true); dv.setUint16(10 + i * 12, 100 + i, true); dv.setUint16(12 + i * 12, p, true); buf.set(body, p); p += body.length; });
+  bodies.forEach((body, i) => { dv.setUint32(4 + i * 12, 0xDEAD, true); dv.setUint16(8 + i * 12, body.length, true); dv.setUint16(10 + i * 12, 100 + i, true); dv.setUint32(12 + i * 12, p, true); buf.set(body, p); p += body.length; });
   return buf;
 }
 function buildMap(parts, rec) {
